# perf: optimize h256 copy_bits fn