--- a/src/tests/tree.rs
+++ b/src/tests/tree.rs
@@ -363,19 +363,14 @@ proptest! {
     }
 
     #[test]
-    fn test_h256_copy_bits(start in 0u8..254u8, size in 1u8..255u8) {
+    fn test_h256_copy_bits(start: u8) {
         let one: H256 = [255u8; 32].into();
-        let target = one.copy_bits(start..(start.saturating_add(size)));
-        for i in start..start.saturating_add(size) {
-            assert_eq!(one.get_bit(i as u8), target.get_bit(i as u8));
+        let target = one.copy_bits(start);
+        for i in start..=core::u8::MAX {
+            assert_eq!(one.get_bit(i), target.get_bit(i));
         }
         for i in 0..start {
-            assert!(!target.get_bit(i as u8));
-        }
-        if let Some(start_i) = start.checked_add(size).and_then(|i| i.checked_add(1)){
-            for i in start_i..=255 {
-                assert!(!target.get_bit(i as u8));
-            }
+            assert!(!target.get_bit(i));
         }
     }