master branch benchmark:
```
SMT generate merkle proof                                                                           
                        time:   [288.30 us 289.57 us 291.47 us]
```

this PR:
```
SMT generate merkle proof                                                                           
                        time:   [191.39 us 193.37 us 194.86 us]
```